test(coppa): extract exactlyAge into a shared helper
Code review noted the timezone-safe exactlyAge helper was copy-pasted verbatim
into three integration test files — a future correction would have to touch all
three, and missing one silently reintroduces the UTC off-by-one bug. Extract it
(plus minorBirthDate/adultBirthDate) into tests/shared/age-helpers.ts and import.
