+++ tests/integration/registration-minor.test.ts
@@ -39,18 +39,8 @@ import { registerRoutes } from '../../packages/api/routes';
 // Helpers
 // ============================================================================
 
-/** Build a YYYY-MM-DD date string for someone who is exactly `years` years old today */
-function exactlyAge(years: number): string {
-  // Use LOCAL date components (not toISOString, which is UTC): calculateAge
-  // parses YYYY-MM-DD as local midnight, so a UTC-formatted date shifts a day at
-  // the boundary and makes an exactly-N-year-old read as N-1.
-  const d = new Date();
-  d.setFullYear(d.getFullYear() - years);
-  const y = d.getFullYear();
-  const m = String(d.getMonth() + 1).padStart(2, '0');
-  const day = String(d.getDate()).padStart(2, '0');
-  return `${y}-${m}-${day}`;
-}
+// Shared timezone-safe date-of-birth helper.
+import { exactlyAge } from '../shared/age-helpers';
 
 /** Build a YYYY-MM-DD date string for someone whose birthday is `days` days from now,
  *  but was born `years` years ago minus those days.
